Trait assembly_simulator::Interface

pub trait Interface: Send + Sync + InterfaceClone {

    fn get_module(&self, _address: &Address) -> Result<Bytecode> { ... }

    fn update_module(
        &self, 
        _address: &Address, 
        _module: &Bytecode
    ) -> Result<()> { ... }

    fn create_module(&self, _module: &Bytecode) -> Result<Address> { ... }

    fn get_data_for(&self, _address: &Address, _key: &str) -> Result<Bytecode> { ... }

    fn print(&self, _message: &str) -> Result<()> { ... }

    fn set_data_for(
        &self, 
        _address: &Address, 
        _key: &str, 
        _value: &Bytecode
    ) -> Result<()> { ... }

    fn get_data(&self, _key: &str) -> Result<Bytecode> { ... }

    fn set_data(&self, _key: &str, _value: &Bytecode) -> Result<()> { ... }

    fn get_time(&self) -> Result<Instant> { ... }

    fn get_random(&self) -> Result<u64> { ... }

    fn module_called(&self) -> Result<()> { ... }

    fn exit_success(&self) -> Result<()> { ... }

    fn get_owned_addresses(&self) -> Result<Vec<Address>> { ... }

    fn get_call_stack(&self) -> Result<Vec<Address>> { ... }

}

Provided methods

fn get_module(&self, _address: &Address) -> Result<Bytecode>

Requires the module in the given address

fn update_module(&self, _address: &Address, _module: &Bytecode) -> Result<()>

Requires to replace the module at the current address

Note: The execution lib will use the current context address for the update module and the new bytecode

fn create_module(&self, _module: &Bytecode) -> Result<Address>

Requires a new address that contains the bytecode sended

fn get_data_for(&self, _address: &Address, _key: &str) -> Result<Bytecode>

Requires the data at the address

fn print(&self, _message: &str) -> Result<()>

Print function for examples

fn set_data_for(
&self,
_address: &Address,
_key: &str,
_value: &Bytecode
) -> Result<()>

Requires to replace the data in the current address

Note: The execution lib will allways use the current context address for the update

fn get_data(&self, _key: &str) -> Result<Bytecode>

fn set_data(&self, _key: &str, _value: &Bytecode) -> Result<()>

fn get_time(&self) -> Result<Instant>

Returns the current time

fn get_random(&self) -> Result<u64>

Returns a random number

fn module_called(&self) -> Result<()>

fn exit_success(&self) -> Result<()>

fn get_owned_addresses(&self) -> Result<Vec<Address>>

fn get_call_stack(&self) -> Result<Vec<Address>>